requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-SocketIO==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0
gevent==23.9.1
gevent-websocket==0.10.1
Werkzeug==3.0.1

# Hardware Communication & Drivers
pyserial==3.5
smbus-cffi==0.5.1
rpi-lgpio==0.4
RPi.GPIO==0.7.1
adafruit-circuitpython-ina219==3.3.10
# Note: MPU6050 library is included in the `modules` directory
rpi-ws281x==5.0.0

# System & Utilities
psutil==5.9.5
numpy==1.26.2

# Computer Vision & ML
# Install opencv via apt: sudo apt-get install python3-opencv
# opencv-python-headless==4.8.1.78
ultralytics==8.0.200

# For PiCamera2 support, install system-wide:
# sudo apt-get install -y python3-picamera2
